+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/trackedentity/TrackedEntityInstanceService.java
@@ -196,7 +196,7 @@ public interface TrackedEntityInstanceService
      * @param entityInstance The to TrackedEntityInstance add.
      * @return A generated unique id of the added {@link TrackedEntityInstance}.
      */
-    int addTrackedEntityInstance( TrackedEntityInstance entityInstance );
+    long addTrackedEntityInstance( TrackedEntityInstance entityInstance );
 
     /**
      * Soft deletes a {@link TrackedEntityInstance}.
@@ -226,7 +226,7 @@ public interface TrackedEntityInstanceService
      * @param id the id of the TrackedEntityInstanceAttribute to return.
      * @return the TrackedEntityInstanceAttribute with the given id
      */
-    TrackedEntityInstance getTrackedEntityInstance( int id );
+    TrackedEntityInstance getTrackedEntityInstance( long id );
 
     /**
      * Returns the {@link TrackedEntityAttribute} with the given UID.
@@ -262,7 +262,7 @@ public interface TrackedEntityInstanceService
      * @param attributeValues    Set of attribute values
      * @return The error code after registering entityInstance
      */
-    int createTrackedEntityInstance( TrackedEntityInstance entityInstance, String representativeId,
+    long createTrackedEntityInstance( TrackedEntityInstance entityInstance, String representativeId,
         Integer relationshipTypeId, Set<TrackedEntityAttributeValue> attributeValues );
 
     List<TrackedEntityInstance> getTrackedEntityInstancesByUid( List<String> uids, User user );
