+++ dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/maintenance/DefaultMaintenanceService.java
@@ -158,7 +158,7 @@ public class DefaultMaintenanceService
     {
         for ( Period period : periodService.getAllPeriods() )
         {
-            int periodId = period.getId();
+            long periodId = period.getId();
             
             try
             {
