+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/trackedentitycomment/DefaultTrackedEntityCommentService.java
@@ -50,7 +50,7 @@ public class DefaultTrackedEntityCommentService
     // -------------------------------------------------------------------------
 
     @Override
-    public int addTrackedEntityComment( TrackedEntityComment comment )
+    public long addTrackedEntityComment( TrackedEntityComment comment )
     {
         commentStore.save( comment );
 
@@ -75,7 +75,7 @@ public class DefaultTrackedEntityCommentService
     }
 
     @Override
-    public TrackedEntityComment getTrackedEntityComment( int id )
+    public TrackedEntityComment getTrackedEntityComment( long id )
     {
         return commentStore.get( id );
     }
