+++ dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/dataset/DataSetStoreTest.java
@@ -83,9 +83,9 @@ public class DataSetStoreTest
         DataSet dataSetB = createDataSet( 'B', periodType );
 
         dataSetStore.save( dataSetA );
-        int idA = dataSetA.getId();
+        long idA = dataSetA.getId();
         dataSetStore.save( dataSetB );
-        int idB = dataSetB.getId();
+        long idB = dataSetB.getId();
 
         dataSetA = dataSetStore.get( idA );
         dataSetB = dataSetStore.get( idB );
@@ -103,7 +103,7 @@ public class DataSetStoreTest
         DataSet dataSet = createDataSet( 'A', periodType );
 
         dataSetStore.save( dataSet );
-        int id = dataSet.getId();
+        long id = dataSet.getId();
 
         dataSet = dataSetStore.get( id );
 
@@ -125,9 +125,9 @@ public class DataSetStoreTest
         DataSet dataSetB = createDataSet( 'B', periodType );
 
         dataSetStore.save( dataSetA );
-        int idA = dataSetA.getId();
+        long idA = dataSetA.getId();
         dataSetStore.save( dataSetB );
-        int idB = dataSetB.getId();
+        long idB = dataSetB.getId();
 
         assertNotNull( dataSetStore.get( idA ) );
         assertNotNull( dataSetStore.get( idB ) );
@@ -145,9 +145,9 @@ public class DataSetStoreTest
         DataSet dataSetB = createDataSet( 'B', periodType );
 
         dataSetStore.save( dataSetA );
-        int idA = dataSetA.getId();
+        long idA = dataSetA.getId();
         dataSetStore.save( dataSetB );
-        int idB = dataSetB.getId();
+        long idB = dataSetB.getId();
 
         assertEquals( dataSetStore.getByName( "DataSetA" ).getId(), idA );
         assertEquals( dataSetStore.getByName( "DataSetB" ).getId(), idB );
