+++ dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/program/ProgramInstanceServiceTest.java
@@ -96,16 +96,16 @@ public class ProgramInstanceServiceTest
 
     private TrackedEntityInstance entityInstanceA;
 
-    private Collection<Integer> orgunitIds;
+    private Collection<Long> orgunitIds;
 
     @Override
     public void setUpTest()
     {
         organisationUnitA = createOrganisationUnit( 'A' );
-        int idA = organisationUnitService.addOrganisationUnit( organisationUnitA );
+        long idA = organisationUnitService.addOrganisationUnit( organisationUnitA );
 
         organisationUnitB = createOrganisationUnit( 'B' );
-        int idB = organisationUnitService.addOrganisationUnit( organisationUnitB );
+        long idB = organisationUnitService.addOrganisationUnit( organisationUnitB );
 
         orgunitIds = new HashSet<>();
         orgunitIds.add( idA );
@@ -172,8 +172,8 @@ public class ProgramInstanceServiceTest
     @Test
     public void testAddProgramInstance()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idB = programInstanceService.addProgramInstance( programInstanceB );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idB = programInstanceService.addProgramInstance( programInstanceB );
 
         assertNotNull( programInstanceService.getProgramInstance( idA ) );
         assertNotNull( programInstanceService.getProgramInstance( idB ) );
@@ -182,8 +182,8 @@ public class ProgramInstanceServiceTest
     @Test
     public void testDeleteProgramInstance()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idB = programInstanceService.addProgramInstance( programInstanceB );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idB = programInstanceService.addProgramInstance( programInstanceB );
 
         assertNotNull( programInstanceService.getProgramInstance( idA ) );
         assertNotNull( programInstanceService.getProgramInstance( idB ) );
@@ -202,7 +202,7 @@ public class ProgramInstanceServiceTest
     @Test
     public void testUpdateProgramInstance()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
 
         assertNotNull( programInstanceService.getProgramInstance( idA ) );
 
@@ -215,8 +215,8 @@ public class ProgramInstanceServiceTest
     @Test
     public void testGetProgramInstanceById()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idB = programInstanceService.addProgramInstance( programInstanceB );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idB = programInstanceService.addProgramInstance( programInstanceB );
 
         assertEquals( programInstanceA, programInstanceService.getProgramInstance( idA ) );
         assertEquals( programInstanceB, programInstanceService.getProgramInstance( idB ) );
@@ -313,8 +313,8 @@ public class ProgramInstanceServiceTest
     @Test
     public void testCompleteProgramInstanceStatus()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idD = programInstanceService.addProgramInstance( programInstanceD );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idD = programInstanceService.addProgramInstance( programInstanceD );
 
         programInstanceService.completeProgramInstanceStatus( programInstanceA );
         programInstanceService.completeProgramInstanceStatus( programInstanceD );
@@ -329,8 +329,8 @@ public class ProgramInstanceServiceTest
         programInstanceA.setStatus( ProgramStatus.COMPLETED );
         programInstanceD.setStatus( ProgramStatus.COMPLETED );
         
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idD = programInstanceService.addProgramInstance( programInstanceD );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idD = programInstanceService.addProgramInstance( programInstanceD );
         
         programInstanceService.incompleteProgramInstanceStatus( programInstanceA );
         programInstanceService.incompleteProgramInstanceStatus( programInstanceD );
@@ -342,8 +342,8 @@ public class ProgramInstanceServiceTest
     @Test
     public void testCancelProgramInstanceStatus()
     {
-        int idA = programInstanceService.addProgramInstance( programInstanceA );
-        int idD = programInstanceService.addProgramInstance( programInstanceD );
+        long idA = programInstanceService.addProgramInstance( programInstanceA );
+        long idD = programInstanceService.addProgramInstance( programInstanceD );
 
         programInstanceService.cancelProgramInstanceStatus( programInstanceA );
         programInstanceService.cancelProgramInstanceStatus( programInstanceD );
